+++ broker/pubsub/subscribe_options.go
@@ -12,25 +12,14 @@ import (
 // Example usages of HandlerFunc could be to log panics or nack/ack messages which cause panics.
 type RecoveryHandler func(context.Context, *pubsub.Message, interface{})
 
-// TODO support more pubsub.ReceiveSettings settings
+type SubscribeOption func(*SubscribeOptions)
+
+// SubscribeOptions TODO support more pubsub.ReceiveSettings settings
 type SubscribeOptions struct {
 	// pubsub ReceiveSettings
 	ReceiveSettings pubsub.ReceiveSettings
 
 	RecoveryHandler RecoveryHandler
-
-	// Other options for implementations of the interface
-	// can be stored in a context
-	Context context.Context
-}
-
-type SubscribeOption func(*SubscribeOptions)
-
-// SubscribeContext set context
-func SubscribeContext(ctx context.Context) SubscribeOption {
-	return func(o *SubscribeOptions) {
-		o.Context = ctx
-	}
 }
 
 // WithRecoveryHandler sets the function for recovering from a panic.