+++ cmd/publish/main.go
@@ -2,6 +2,9 @@ package main
 
 import (
 	"context"
+	"os"
+	"os/signal"
+	"syscall"
 	"time"
 
 	"cloud.google.com/go/pubsub"
@@ -12,8 +15,8 @@ import (
 )
 
 func main() {
-	// broker.DefaultBroker = broker.NewBroker(context.Background(), broker.ProjectID("my-project-id")); // use cfg.pubsub.ProjectID
-	broker.DefaultBroker = broker.NewBroker(context.Background())
+	appCtx, stop := signal.NotifyContext(context.Background(), syscall.SIGHUP, syscall.SIGTERM, syscall.SIGINT, os.Interrupt)
+	broker.DefaultBroker = broker.NewBroker(appCtx, broker.ProjectID("my-project-id"))
 
 	msg := pubsub.Message{
 		ID:         uuid.New().String(),
@@ -25,12 +28,13 @@ func main() {
 	var pub broker.Publisher
 	if pub, err = broker.NewPublisher("toolkit-in-dev",
 		broker.WithPublishSettings(pubsub.PublishSettings{Timeout: time.Second * 5})); err != nil {
-		log.Error().Err(err).Send()
+		log.Fatal().Err(err).Msg("Failed to create publisher for topic: toolkit-in-dev")
 	}
 
 	if err := pub.Publish(context.Background(), &msg); err != nil {
-		log.Error().Err(err).Send()
+		log.Error().Err(err).Msg("Failed to publish to topic: toolkit-in-dev")
 	}
 
-	pub.Stop()
+	stop()
+	log.Info().Msg("Shutting down gracefully, press Ctrl+C again to force")
 }