+++ util/signals/signal_windows.go
@@ -1,9 +1,0 @@
-// +build windows
-
-package signals
-
-import (
-	"os"
-)
-
-var shutdownSignals = []os.Signal{os.Interrupt}